+++ platform/web.js
@@ -3,3 +3,6 @@ module.exports.Mouse = require('../lib/mouse');
 module.exports.Math = require('../lib/util/math');
 require('../lib/tween');
 require('../lib/loader.web');
+
+module.exports._extend = require('../lib/util/extend');
+module.exports._create = require('../lib/util/create');